podoop and a dot
for now the main character is going to be called podoop.
let's see how many times I misspell that before I change it.

the plan is to dynamically draw the tentacles between an anchor point on
the head and the (moveable) tip of the tentacle. for now the tip is just
moveable