write the tentacle movement to a sprite
new strategy - instead of using an algorithm to draw the tentacle
connecting the tip and the anchor, allow free drawing and write the
results to memory so that the entire sprite can be cleared/moved in a
couple actions. this also allows for free tentacle movement.

the problems to solve still are:
  backtracking
  max length
  moving outside the sprite limit