allow backtracking when moving the tentacle
this seems a bit overly complex - there's a lot of bitwise operations
going on to NAND the about-to-be-written pixel with the current state of
the tentacle sprite in order to determine if there's a collision.
normally this could be done by drawing to the screen and checking vf but
that would produce flicker. it also wouldn't let us draw over other
objects in the same plane but I'm not sure I wan't to allow that anyway.

when a collision is detected the current algorithm assumes it's
backtracking and will erase the previous position of the tentacle, but
this fails horribly with any other collision case.