primitive food collection
it's important to not end the game when collecting food, and it's
important to actuall collect the food.

in order to tell the difference between a game-ending collision and a
food gather collision I redraw the food on the board. if there's a
collision with the food then the first redraw (erasing the food) will
leave a pixel behind where the arm has collided with it. redrawing the
food again (to put it back) will then also set vf to 1 because of the
pixel left behind. as long as food is never drawn on top of an existing
arm this should work just fine. however, it only works for the shrimp.