draw food from a board in memory
right now food and collision detection is just done really basically and
only works for one sepcific positioning of one type of food. this is the
first step in making that more flexible and will allow for making full
levels.

a board of food is stored in memory in sets of 4 bytes. each set
represents a piece of food and each byte represents information
about that piece (it's type, whether or not it's been collected, and its
coordinates). the draw-board subroutine loops through these sets and
draws the food to the screen by looking up the appropriate sprite in the
food-table.