extend new food system to collection
the new collection system works very similarly to the old one, this time
it just loops through each of the foods. it's really important that the
collision detection draws are done consecutively to minimize flicker.

unfortunately, this loop occurs for every movement so it could get
really costly. I'd like to short circuit out of it as soon as I can.