combine the starting position and food table into a level
the goal of the game will be to progress through a series of
increasingly challenging levels. each level will have different food
starting positions and a different food table, as well as background
sprites and potentially other aspects. when making a new level I
want to be able to just invoke a "make level" subroutine that will set
everything up based on the level information.