make level 3
now we're starting to roll! I noticed that I needed to clear the food
board if a level had more food pieces than any of its successors. I
decided to make this the main clearing action as well and to do this
whenever switching levels so each level creating subroutine doesn't have
to call it. there's already enough repetition in them anyway.