difficulty modes! easy, normal, and hard
I've commented out all of the multiplayer code because I'm definitely
not getting to it but I'd like that infrastructure to remain there. In
the meantime, I've replaced the game mode select with a blinking 'press
to start' which feels very arcadey.

In order to implement difficulty levels I've bumped the clock rate up to
60 cycles/frame. During the main game loop I set a delay on when the
arms move and detect for collisions. In hard mode the delay is only
1/60th of a second, so it's essentially moving every frame. I can't beat
most of the levels at this speed. Normal slows it down to every 4
frames, and easy to every 10. Normal could probably be a bit faster but
it feels about what it did before.

The code is starting to get a biy messy here becuase it's T minus 3
hours until deadline. Sorry future jackie.