give the player a reward for making it to the end
I ran into a slight problem where the animation just stopped playing
because the program was getting too long. I had to delete the old
puzzle/multiplayer text to get it to work. I'm not 100% sure why this is
the case.

Also, I really don't like the fact that the animation uses full 16x16
SUPERCHIP sprites but to get it to draw cleanly that's the only way.